Adding a 'filter' parameter to limit what will be synced.
The `filter` parameter is optional, but it (or something like it) is needed to limit what is pulled down from S3. By default the `sync` operation syncs the working directory which, for us, also syncs the `scripts` folder. This could result in some squirrely behavior if we `get` the last `put`'d script directory.